feat : 최대 구입 가능 금액 검증 기능 추가
최대 구입 한도인 10만원을 초과하지 못하게 하였음

+++ src/main/java/model/Amount.java
@@ -1,14 +1,13 @@
 package model;
 
-import java.math.BigInteger;
 
 public class Amount {
-    private final BigInteger purchaseAmount;
-    public Amount(BigInteger purchaseAmount){
+    private final int purchaseAmount;
+    public Amount(int purchaseAmount){
         this.purchaseAmount = purchaseAmount;
     }
 
-    public BigInteger getPurchaseAmount() {
+    public int getPurchaseAmount() {
         return purchaseAmount;
     }
 }
